Class refactored.
