+++ vcf-dumper/vcf-dumper-ws/src/main/java/uk/ac/ebi/eva/vcfdump/server/HtsGetError.java
@@ -1,0 +1,28 @@
+package uk.ac.ebi.eva.vcfdump.server;
+
+class HtsGetError {
+    private String error;
+
+    private String message;
+
+    HtsGetError(String error, String message) {
+        this.error = error;
+        this.message = message;
+    }
+
+    public String getError() {
+        return error;
+    }
+
+    public void setError(String error) {
+        this.error = error;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
